feat: renamed ViewProfile and added its ViewModel
