+++ lib/app/app.router.dart
@@ -46,7 +46,7 @@ import '../ui/view/saved_items/saved_items_view.dart';
 import '../ui/view/set_status/set_status_view.dart';
 import '../ui/view/sign_up/sign_up_view.dart';
 import '../ui/view/splashscreen/splashscreen.dart';
-import '../ui/view/view_profile_page/view_profile_page.dart';
+import '../ui/view/view_profile_page/view_profile.dart';
 import '../ui/view/workspace/add_workspace/add_workspace_view.dart';
 import '../ui/view/workspace/create_workspace/create_workspace.dart';
 import '../ui/view/workspace/workspace_different_email/difference_email_workspace_view.dart';
@@ -171,7 +171,7 @@ class StackedRouter extends RouterBase {
     RouteDef(Routes.addPluginView, page: AddPluginView),
     RouteDef(Routes.useDifferentEmailView, page: UseDifferentEmailView),
     RouteDef(Routes.editPluginView, page: EditPluginView),
-    RouteDef(Routes.viewProfilePage, page: ViewProfilePage),
+    RouteDef(Routes.viewProfilePage, page: ViewProfile),
     RouteDef(Routes.setStatusView, page: SetStatusView),
     RouteDef(Routes.profilePageView, page: ProfilePageView),
     RouteDef(Routes.preferenceView, page: PreferenceView),
@@ -346,12 +346,12 @@ class StackedRouter extends RouterBase {
         settings: data,
       );
     },
-    ViewProfilePage: (data) {
+    ViewProfile: (data) {
       var args = data.getArgs<ViewProfilePageArguments>(
         orElse: () => ViewProfilePageArguments(),
       );
       return MaterialPageRoute<dynamic>(
-        builder: (context) => ViewProfilePage(
+        builder: (context) => ViewProfile(
           key: args.key,
           isActive: args.isActive,
         ),
