+++ lib/ui/view/view_profile_page/view_profile_viewmodel.dart
@@ -1,0 +1,3 @@
+import 'package:stacked/stacked.dart';
+
+class ViewProfileViewModel extends BaseViewModel {}
